requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.6
celery[redis]==5.2.7
pandas==2.0.0
openpyxl==3.1.2
